refactor: extract PR management and webhook forwarder from daemon/mod.rs
Extract two more domain-specific modules:

- pr.rs (~1,500 lines): PR polling, auto-merge, reviewer spawning,
  review spawn queue processing, PR comment nudging, and related helpers
- webhook_fwd.rs (~260 lines): GitHub CLI webhook forwarder watchdog,
  extension management, and stale hook cleanup

Updates call sites in mod.rs, snapshot.rs. Adds helpers::* import to
test module for functions that moved to helpers.rs in the prior PR.

mod.rs shrinks from ~4,060 to ~2,300 lines.
